Adds amulet test - ceph.conf alternatives removal
Test checks for the removal of ceph.conf alternatives when ceph-mon
relation is broken. If ceph.conf is not present when the relation is
still joined, the test fails. If ceph.conf is not removed after the
relation is broken, the test fails. The test will pass if an existing
ceph.conf file is removed after the ceph-mon relation is broken.

Change-Id: I3b348a58bd2e3ebbbecbd3bbb4307c490a0c4ea4
Related-Bug: 1778084