Require relation to nova-compute application
Without a relation to at least one nova-compute application a
cinder-ceph backend will not be functional as the libvirt
secrets will not have been created to allow access to the
ceph cluster from libvirt/qemu.

Add a simple context to check that the 'ceph-access' relation
is present.  This will result in a blocked status if the
relation is not detected - for example:

  Missing relations: nova-compute

Change-Id: Iedbf4aafc2348cbf6f14257417e86aa9aeb48a81
Closes-Bug: 1718051